added test project for desktop
